+++ modules/config/configuration.py
@@ -2,5 +2,9 @@
 # it doesnt exist anymore but i dont want to remove from gitignore to be safe
 
 import modules.config.config_tools as conftools
+import modules.versioning_tools as versioning_tools
 
 config = conftools.Config.load()
+
+remote_version = versioning_tools.get_remote_version()
+local_version = versioning_tools.get_local_version()
