+++ modules/notifications/ntfy.py
@@ -2,6 +2,7 @@ import requests
 from modules.colors.ansi_codes import RESET, RED, GREEN, BLUE, YELLOW, WHITE, PURPLE, CYAN, LIGHT_CYAN, SUPER_LIGHT_CYAN, ORANGE, ansi_is_supported
 from modules.config.configuration import config
 
+
 def send_notification(text, shorturl):
     headers={
         "X-Click": shorturl, # notification click action
